Better support of multiple positions
We cannot pass a position as parameter, as this changes the configuration of the toaster (toast
container), so it changes future displays.
Instead, we have to make one instance of container per wanted position, which is set in the config
options. Therefore, we have to define also a toasterId option, so that calls to pop() target one
toaster or the other.
Of course, retro-compatibility is maintained: toaster without defined toasterId answer to calls
without id set (but not to those targeting a specific toaster).