Events were not being setup on reintroduced child scope
It is my belief that the directive should be able to be attached on any scope,
and that the event factory should needs to know to set itself up again, since
the subscriptions to those events are added when the directive is compiled.

Test case:
Directive is on a child scope.
Remove this scope.
Re attach the scope.
The event factory does not run, so the events are not attached.