#153 - Pass Data To Directive
Added a new feature to allow data to be passed to a directive
responsible for loading a template.  A new toast parameter,
directiveData, will bind data directly to the compile directive
template's scope, allowing the caller to provide data to a toast
template when invoking the toast.

Additional error handling was added to the directiveTemplate directive
to support this change as well as new test coverage.